Dart 2 and stronger mode

+++ serializer/lib/src/codecs/date_time_milliseconds.dart
@@ -7,7 +7,7 @@ import 'type_codec.dart';
 /// A milliseconds since epoch DateTime codec.
 class DateTimeMillisecondsSinceEpochCodec extends TypeCodec<DateTime> {
   DateTime decode(dynamic value, {Serializer serializer}) =>
-      new DateTime.fromMillisecondsSinceEpoch(value);
+      new DateTime.fromMillisecondsSinceEpoch(value as int);
   dynamic encode(DateTime value,
           {Serializer serializer, bool useTypeInfo, bool withTypeInfo}) =>
       value.millisecondsSinceEpoch;
